+++ src/polyfills.js
@@ -2,5 +2,6 @@
 // TODO: These are better included using webpack.config.js
 // But that doesn't seem to work with Babel at the moment.
 import 'core-js/modules/es6.number.is-safe-integer';
+import 'core-js/modules/es6.number.is-integer';
 import 'core-js/modules/es6.number.is-finite';
 import 'core-js/modules/es6.number.is-nan';